Add automatic filling of the currently selected range by hitting enter
When the user hits enter, the software tries to guess what the user was going to put into the currently selected range.
Three different logics are implemented: one to fill the sequence, one to fill the pattern and one to fill the fx-track.
The fx-track is the easiest: it interpolates between the two ends of the selected range.
The pattern is filled by finding the last line with any notes, and then looping the notes to fill the selected range.
The sequencer filling logic is most complicated; 1) With a single pattern, it repeat it. 2) With two pattern, it
tries to extrapolate a sequence. [1,2,,,,,] becomes [1,2,3,4,5,6,7]. 3) With more patterns, they are repeated. i.e.
[1,2,1,3,,,] becomes [1,2,1,3,1,2,1]. The behaviour is also described in the help.html.

Fixes #18.