Add undo functionality
Ctrl-z is undo, while alt-ctrl-z is redo. Before the song is modified, pushes a deep copy of the whole
mSong into a stack. The deep copying takes of the order of 20 ms on an oldish computer
so this does not seem to be a bottle neck. The undo pushes are throttled: the undo state is pushed
at maximum once per 3 seconds, limiting the number of deep copies generated. The stack is also
limited to a maximum of 100 items.